Updating private api controller to meet api requirements
Now it uses the various util and exception classes for less coupling and better testing.

+++ back-end/src/main/java/org/talentboost/finaltask/backend/controllers/PrivateAPIController.java
@@ -1,18 +1,18 @@
 package org.talentboost.finaltask.backend.controllers;
 
-import org.apache.catalina.connector.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.talentboost.finaltask.backend.data.Image;
+import org.talentboost.finaltask.backend.exceptions.UnsupportedMimeTypeException;
+import org.talentboost.finaltask.backend.exceptions.VladoException;
 import org.talentboost.finaltask.backend.repository.ImageRepository;
+import org.talentboost.finaltask.backend.util.FileUtils;
 
+import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.*;
 
 @RestController
@@ -21,50 +21,40 @@ public class PrivateAPIController {
     @Autowired
     private ImageRepository repository;
 
-    private final String STATIC_FOLDER = "../static";
-    private final Map<String, String> mimeTypeMappings;
+    private final FileUtils fileUtils;
+    private final String BASE_URL = "http://localhost:3333/";
 
     public PrivateAPIController() {
         Map<String, String> tmpMap = new HashMap<>();
+        String staticFolder = "../static";
 
         tmpMap.put("image/jpeg", "jpg");
         tmpMap.put("image/png", "png");
         tmpMap.put("image/gif", "gif");
 
-        mimeTypeMappings = Collections.unmodifiableMap(tmpMap);
+        this.fileUtils = new FileUtils(staticFolder, Collections.unmodifiableMap(tmpMap));
+    }
+
+    @ExceptionHandler
+    public void handleUnsupportedMimeType(
+            VladoException e,
+            HttpServletResponse response
+    ) throws IOException {
+        response.sendError(HttpStatus.BAD_REQUEST.value(), e.getMessage());
     }
 
     @PostMapping("/meme")
     public ResponseEntity<Image> createImage(
             @RequestParam("title") String title,
             @RequestParam("file") MultipartFile[] files
-    ) {
-        String uuid = UUID.randomUUID().toString();
+    ) throws IOException {
+        String fileName = fileUtils.createFile(files[0]);
 
-        MultipartFile file = files[0];
-        String fileMimeType = Objects.requireNonNull(file.getContentType());
-        String fileExtension = mimeTypeMappings.get(fileMimeType);
-
-        if (fileExtension == null) {
-            return ResponseEntity.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE).build();
-        }
-
-        String fileName = uuid + "." + fileExtension;
-
-        try (
-                InputStream fileInputStream = file.getInputStream()
-        ) {
-            Files.copy(
-                    fileInputStream,
-                    Paths.get(STATIC_FOLDER + "/" + fileName)
-            );
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        Image toBeSaved = new Image(title, BASE_URL + fileName);
 
         return ResponseEntity
                 .status(HttpStatus.CREATED)
-                .body(repository.save(new Image(title, "http://localhost:3333/" + fileName)));
+                .body(repository.save(toBeSaved));
     }
 
     @PutMapping("/meme/{id}")
@@ -72,21 +62,37 @@ public class PrivateAPIController {
             @PathVariable("id") int id,
             @RequestParam("title") String title,
             @RequestParam("file") MultipartFile[] files
-
-    ) {
+    ) throws IOException {
         Optional<Image> toAlter = repository.findById(id);
-        toAlter.ifPresent(image -> {
-            // Logic
+        if(toAlter.isPresent()) {
+            Image image = toAlter.get();
+
+            image.setTitle(Objects.requireNonNull(title));
+
+            if(files != null && files.length > 0) {
+                String oldFileName = fileUtils.getFileNameFromUrl(image.getUrl());
+                String newFileName = fileUtils.createFile(files[0]);
+
+                fileUtils.deleteFile(oldFileName);
+                image.setUrl(BASE_URL + newFileName);
+            }
+
             repository.save(image);
-        });
+        }
         return toAlter.orElse(null);
     }
 
     @DeleteMapping("/meme/{id}")
     public void deleteImage(
             @PathVariable("id") int id
-    ) {
-        repository.deleteById(id);
+    ) throws IOException {
+        Optional<Image> toBeDeleted = repository.findById(id);
+        if(toBeDeleted.isPresent()) {
+            String url = toBeDeleted.get().getUrl();
+
+            fileUtils.deleteFile(fileUtils.getFileNameFromUrl(url));
+            repository.deleteById(id);
+        }
     }
 
 }
